vocab_dictionnary/requirements.txt
djangorestframework
markdown
django-filter
psycopg2-binary
django-rest-swagger
Pillow
djangorestframework-simplejwt